+++ src/sets.rs
@@ -4,7 +4,7 @@
 //  Created:
 //    13 Jan 2025, 15:26:24
 //  Last edited:
-//    15 Jan 2025, 18:02:44
+//    16 Jan 2025, 12:16:02
 //  Auto updated?
 //    Yes
 //
@@ -46,9 +46,9 @@ pub type Statements = MapAsync<Arc<Message>>;
 /// Defines a _synchronous_ set for keeping track of the (current) time.
 pub struct Times {
     /// The current time, if any.
-    current: Option<u128>,
+    current: Option<u64>,
     /// The set of all known times.
-    times:   HashSet<u128>,
+    times:   HashSet<u64>,
 }
 impl Times {
     /// Creates a new, empty Times.
@@ -58,27 +58,27 @@ impl Times {
     #[inline]
     pub fn new() -> Self { Self { current: None, times: HashSet::new() } }
 }
-impl justact::Set<u128> for Times {
+impl justact::Set<u64> for Times {
     type Error = Infallible;
 
     #[inline]
-    fn get(&self, id: &u128) -> Result<Option<&u128>, Self::Error> { Ok(self.times.get(id)) }
+    fn get(&self, id: &u64) -> Result<Option<&u64>, Self::Error> { Ok(self.times.get(id)) }
 
     #[inline]
-    fn iter<'s>(&'s self) -> Result<impl Iterator<Item = &'s u128>, Self::Error>
+    fn iter<'s>(&'s self) -> Result<impl Iterator<Item = &'s u64>, Self::Error>
     where
         u128: 's,
     {
         Ok(self.times.iter())
     }
 }
-impl justact::SetSync<u128> for Times {
+impl justact::SetSync<u64> for Times {
     #[inline]
-    fn add(&mut self, elem: u128) -> Result<bool, Self::Error> { Ok(self.times.insert(elem)) }
+    fn add(&mut self, elem: u64) -> Result<bool, Self::Error> { Ok(self.times.insert(elem)) }
 }
 impl justact::Times for Times {
-    type Subset = Option<u128>;
-    type Timestamp = u128;
+    type Subset = Option<u64>;
+    type Timestamp = u64;
 
     #[inline]
     fn current(&self) -> Result<Self::Subset, Self::Error> { Ok(self.current) }
@@ -87,7 +87,7 @@ impl justact::TimesSync for Times {
     #[inline]
     fn add_current(&mut self, timestamp: Self::Timestamp) -> Result<bool, Self::Error> {
         // Always add to the set
-        <Self as justact::SetSync<u128>>::add(self, timestamp)?;
+        <Self as justact::SetSync<u64>>::add(self, timestamp)?;
 
         // Then add it as the current timestamp, but only by checking if it exists already
         let existed: bool = if let Some(current) = self.current { current == timestamp } else { false };
